refactor: organize sync integrations into src/integrations/
Move 30 sync integration files from the flat src/ directory into
organized subdirectories under src/integrations/:

  integrations/oura/       — client, sync, process, webhook-api, webhook-router
  integrations/garmin/     — client, sync, process, types
  integrations/lastfm/     — client, sync, session-grouping
  integrations/rescuetime/ — client, sync
  integrations/ical/       — sync
  integrations/activitywatch/ — sync
  integrations/owntracks/  — router

Files renamed to shorter names within their directories (e.g.
oura-sync.ts → integrations/oura/sync.ts). All imports updated
across api.ts, services, mcp tools, and cross-references.

Closes #590
