+++ apps/backend/src/api.ts
@@ -16,9 +16,8 @@ import cors from 'cors'
 import express, { json, type NextFunction, type Request, type RequestHandler, type Response } from 'express'
 import { Client } from 'pg'
 
-import type { OuraDataType } from './oura-sync.ts'
+import type { OuraDataType } from './integrations/oura/sync.ts'
 
-import { processActivityWatchEvents } from './activitywatch-sync.ts'
 import { createAuth } from './auth.ts'
 import {
   ackOutboundSync,
@@ -48,18 +47,18 @@ import {
   updateDetectedLocation,
   upsertUserSettings,
 } from './db/index.ts'
-import { processActivityDetail } from './garmin-process.ts'
-import { syncAllGarminData } from './garmin-sync.ts'
-import { garminClient } from './garmin.ts'
 import { httpError, isHttpError } from './http-error.ts'
-import { syncAllCalendars } from './ical-sync.ts'
-import { syncLastFmData } from './lastfm-sync.ts'
+import { processActivityWatchEvents } from './integrations/activitywatch/sync.ts'
+import { garminClient } from './integrations/garmin/client.ts'
+import { processActivityDetail } from './integrations/garmin/process.ts'
+import { syncAllGarminData } from './integrations/garmin/sync.ts'
+import { syncAllCalendars } from './integrations/ical/sync.ts'
+import { syncLastFmData } from './integrations/lastfm/sync.ts'
+import { ouraClient } from './integrations/oura/client.ts'
+import { syncAllOuraData, syncOuraDataType } from './integrations/oura/sync.ts'
+import { createOwnTracksRouter } from './integrations/owntracks/router.ts'
+import { syncRescueTimeData } from './integrations/rescuetime/sync.ts'
 import { createMcpRouter } from './mcp.ts'
-import { createScrobblesRouter } from './routes/scrobbles-router.ts'
-import { syncAllOuraData, syncOuraDataType } from './oura-sync.ts'
-import { ouraClient } from './oura.ts'
-import { createOwnTracksRouter } from './owntracks.ts'
-import { syncRescueTimeData } from './rescuetime-sync.ts'
 import { createActivitiesRouter } from './routes/activities-router.ts'
 import { createActivityTypesRouter } from './routes/activity-types-router.ts'
 import { createAdminRouter } from './routes/admin-router.ts'
@@ -77,6 +76,7 @@ import { createNotesRouter } from './routes/notes-router.ts'
 import { createOAuthRouter } from './routes/oauth-router.ts'
 import { createReportsRouter } from './routes/reports-router.ts'
 import { createScreentimeCategoriesRouter } from './routes/screentime-categories-router.ts'
+import { createScrobblesRouter } from './routes/scrobbles-router.ts'
 import { createSettingsRouter } from './routes/settings-router.ts'
 // tags-router removed: tags are now activities
 import { createTrainingLoadRouter } from './routes/training-load-router.ts'
