+++ apps/backend/src/integrations/garmin/types.d.ts
