+++ apps/backend/src/integrations/ical/sync.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, test, vi } from 'vitest'
 
-import * as db from './db/index.ts'
-import { parseICalEvents, syncAllCalendars, syncCalendar } from './ical-sync.ts'
+import * as db from '../../db/index.ts'
+import { parseICalEvents, syncAllCalendars, syncCalendar } from './sync.ts'
 
 vi.mock('./db', () => ({
   getAllSyncStates: vi.fn().mockResolvedValue([]),
