+++ apps/backend/src/integrations/lastfm/client.test.ts
@@ -5,7 +5,7 @@
 import axios from 'axios'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { lastfmClient } from './lastfm.ts'
+import { lastfmClient } from './client.ts'
 
 vi.mock('axios')
