+++ apps/backend/src/integrations/oura/sync.ts
@@ -9,19 +9,14 @@
 
 import { addMinutes, isFuture, subDays } from 'date-fns'
 
-import type { ouraClient } from './oura.ts'
+import type { ouraClient } from './client.ts'
 
-import { getSyncState, getUserSettings, type SyncState, upsertSyncState } from './db/index.ts'
-import { type OuraDataType, processOuraData } from './oura-process.ts'
-import { triggerCalorieComputation } from './services/calorie-computation.ts'
+import { getSyncState, getUserSettings, type SyncState, upsertSyncState } from '../../db/index.ts'
+import { triggerCalorieComputation } from '../../services/calorie-computation.ts'
+import { type OuraDataType, processOuraData } from './process.ts'
 
 // Re-export for consumers that import from oura-sync
-export {
-  computeSleepMinutes,
-  convertOuraSleepPhases,
-  processOuraData,
-  type OuraDataType,
-} from './oura-process.ts'
+export { computeSleepMinutes, convertOuraSleepPhases, processOuraData, type OuraDataType } from './process.ts'
 
 /** Default start date for historical sync (90 days back) */
 const DEFAULT_SYNC_HISTORY_DAYS = 90
