+++ apps/backend/src/services/sync-provider.ts
@@ -7,8 +7,8 @@
 
 import { isBefore, subMinutes } from 'date-fns'
 
-import type { GarminClient } from '../garmin.ts'
-import type { ouraClient } from '../oura.ts'
+import type { GarminClient } from '../integrations/garmin/client.ts'
+import type { ouraClient } from '../integrations/oura/client.ts'
 import type { SyncProvider } from './queries.ts'
 
 import { getSyncState } from '../db/index.ts'
@@ -17,15 +17,19 @@ import {
   isRateLimited as isGarminRateLimited,
   syncActivityDetails,
   syncGarminDataType,
-} from '../garmin-sync.ts'
-import { syncAllCalendars } from '../ical-sync.ts'
-import { syncLastFmData } from '../lastfm-sync.ts'
-import { isRateLimited as isOuraRateLimited, type OuraDataType, syncOuraDataType } from '../oura-sync.ts'
+} from '../integrations/garmin/sync.ts'
+import { syncAllCalendars } from '../integrations/ical/sync.ts'
+import { syncLastFmData } from '../integrations/lastfm/sync.ts'
+import {
+  isRateLimited as isOuraRateLimited,
+  type OuraDataType,
+  syncOuraDataType,
+} from '../integrations/oura/sync.ts'
 import {
   isRateLimited as isRescueTimeRateLimited,
   needsSync as rescueTimeNeedsSync,
   syncRescueTimeData,
-} from '../rescuetime-sync.ts'
+} from '../integrations/rescuetime/sync.ts'
 import { auditError, auditInfo, auditWarn } from './audit-log.ts'
 import { getSettings } from './settings.ts'
